Fix goblin.created event sending race condition
The bug appears when the same goblin is created quickly several times.
The first creation is not finished, the counter is greater than 1. In
this case, the created event is never sent.

The fix is simple. A notified bool is used in order to detect the first
creation even if the counter is > 1.