Fix race when a same goblin is creating multiple times
It happens when doQuest is called and the forst doQuest for create was
not ended. The ref counter was greater than 1 and the event was never
sent by the first `create` call.

The fix is using a flag `created` in order to accept ref count > 1. Then
`created` must be false in order to send event. It's possible to have
the ref counter greater than 1.