Propagate the feed when creating new goblins
The feed is used by the goblin.created event in order to add the
new branch only if it's matching the right feed. The previous way
was working only when a bus client is created for each laboratory
because the orcName is used for filtering the goblin.created events.
In the case of zepplin, only one bus client exists but a lot of
laboratories (carnotzets) are created. The new branches must be
added only for the appropriate feed and not accordingly to the
current orcName.