Fix 'xxx' error in warehouse by acquiring the mutex ASAP
The 'xxx' error was a strange error when a create was running and a
state was already available in the warehouse and it makes no sense. When
a state already exists, the goblin is already created or a create is
running. The bug was just because the test was not in the creation
mutex. It was possible to have more than one create at a time and it's
not acceptable.