Add TTL support with a RankedCache
When a goblin is in the cache, its TTL is set to Infinity (the real
Infinity value (float)). When a goblin is out of the cache (search for
'out' in this commit, then the TTL is restored to a value in the
warehouse. The TTL is used by the collector in order to decide when a
branch must be killed or not event if the list of owners (createdBy) is
empty.