Use killCreatedBy instead of removeCreatedBy where appropriate
quest.kill must update the subscriptions when an owner is removed.
It was already the case with collect when a branch is removed but
in the case where the branch stays alive, it's necessary to clean
the subscriptions too and it's more efficient to do this stuff
directly with killCreatedBy instead by a new collect stuff.