Impl. queue and worker queue service builder
You can now use this API for exporting a queue in a
service file <queue-name.js>:

Goblin.buildQueue(<queue-name>,{
  queueSize: <1..n>,
  sub: '*::*.topic',
  jobIdGetter: (msg) => msg.data.entityId,
})

queueSize is optional and will be set to 100 if not provided.
jobIdGetter is optional, msg.id will be used if not provided.

Goblin.buildQueueWorker(<queue-name>,{
 workQuest: function*(quest){...}
})

important: use the same queue name for the worker,
and name your service file: <queue-name>-worker.js

See goblin-tests for example service file and usage.