Don't unlink the guildEnforcer database anymore
When two master app with the same ID are running, deleting the guild
enforcer cache is a problem on Windows where the file is blocked EBUSY.
This commit prefer a SQL DELETE of the content in order to prevent bad
behaviours on Windows. Note that this database makes sens only for
tribes and usual desktop apps are not using tribes. It's safe to delete
the whole content at each start of the same app.