Fix enroleUser when multiple claims exists in a token data
For example in the case of an app which defines a special claim:

tokenData.myclaim = 'admin'
tokenData.aud = 'mysite.com'

The previous implementation overload the rank accordingly to the
last claim detected via the for.of loop. Sometimes it's possible
that 'myclaim' is after 'aud', or 'aud' is after 'myclaim'.

Here, the fix is simple. The 'aud' claim is used only as fallback
when the special claim is not found.