Load the persisted state of Archetype logic states on create
On create, the Elf tries to load the state from the last persist action
available in the Cryo store. It means that once the Elf is created, it's
possible to read immediatly the last state.

Note that in this case, the user's create quets is not called. It's
called only in the case where no persist action exists for the specified
ID.