Fix major bug with the state when it's shared
The state was shared between the reducers and the quests and it's wrong.
When it's shared, the state is not refreshed where appropriate when the
state ref is passed somewhere else. For example:

  const {state} = this;

was not like

  this.state

because the second way uses a getter on this. This getter was using the
state provided by the logic stuff. Now the state is a new one with its
own ref.