Fix actions order with the persisted actions returned by the server
By following the strict order, we are sure that the last persisted
action replayed on the client, uses the last commitId known by the
server. This commit fixes strange behaviouris where a same database was
synchronized multiple times for no good reason, because the actions were
already persisted on the client side. The problem was only that the last
commitId on the client side was not at the appropriate location.

The fix is mostly one major change. Instead of using a plain object with
the goblin's id as keys, an array is used instead. The array ensures
that we preverse the order. A comment (section (1)) in this commit gives
a simple example on how it works.

These changes break the exchanges between the server and the clients.