Fix replaying with insertOrCreate/Replace helpers
Without this fix, it's possible to lose data because in this case, the
create action is not generated and only diff actions are sent to the
server. When the goblin is new, "diff" are applied on an empty state.

Now, insertOrCreate and insertOrReplace use their own logic handler
which works like the $create logic handler. It's a sort of first create
with the whole payload.