Wrap the whole actions in only one big transaction
Then it's possible to rollback in case of shutdown or in case of error.
Note that here we lose potential other actions sent by persist when the
user is working if a rollback occures. Of course, it's always better to
be sure that no more sync are running first.

TODO: the last chance dialog should ask the user if he's really sur to
quit, in the case where some hot work can exist in the transactions.