Prevent empty range in the case of fromId doesn't exists anymore
It's possible that a commit is collected and then, not known by the
server. Most of time Yeti bootstraps again the database at the statup.
But when Yeti is in offline mode, the bootstrap is no longer possible.
In this case, when the last commitId is unknown, we must return a full
range (it's a bit like a bootstrap but with runtime support).