Fix a regression with actions that were sent while an interrupted sync occures
These actions must not be applied before the ripleyServer call. Because
in this case, a newer lastCommitId is used in order to retrieve the new
actions from the server.
The zero actions that are on the server can be returned by ripleyServer
like any other actions. The fix is simple, just stop the apply, let's
ripleyServer does the usual job. The replace of zero commitId must be
donc only at the end like all other actions.