Use asc order for the extracted actions
The reversed order is no longer necessary because it was used (in much
older code) for unshifting actions in the final persited list. Now that
we use a temporary _persisted list, we can push one by one in the final
order. We must (just) ensure that the tempoery _persisted list is sent
to the client before the user's persisted values. it's for this reason
that the final persisted list is the temporary _persisted list concated
to the user persisted list.

  persisted = _persisted + persisted